ObjectDetectionDemo/requirements.txt
ultralytics 
opencv-python
numpy
pillow 
tkinter